added option to input parameters ranges to shift, moved parser to io.py, added tests
